% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCarbonIntensity.R
\name{CalculateCarbonIntensity}
\alias{CalculateCarbonIntensity}
\title{Main function to implement the calculation of carbon intensity.}
\usage{
CalculateCarbonIntensity(L)
}
\arguments{
\item{L}{A list containing data from preprocessing supplied optional input
files returned by the processModuleInputs function. This list has two
components: Errors and Data.}
}
\value{
A list that is the same as the input list with an additional
Warnings component.
}
\description{
\code{CalculateCarbonIntensity} manages the calculation of fuel and
electricity carbon intensities by transportation mode.
}
\details{
This function manages the calculation of average fuel carbon intensity for
each modeled transportation mode. Household vehicle, commercial service
vehicle, car service vehicle, and heavy truck average fuel carbon intensities
are calculated at the region level. Transit van, transit bus, and transit
rail average fuel carbon intensities are calculated at the Marea level.
Electricity carbon intensity is calculated at the Azone level. This function
calls functions which calculate carbon intensity for each mode.
}
