% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCarbonIntensity.R
\docType{data}
\name{CalculateCarbonIntensitySpecifications}
\alias{CalculateCarbonIntensitySpecifications}
\title{Specifications list for CalculateCarbonIntensity module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CalculateCarbonIntensity.R script.
}
\usage{
CalculateCarbonIntensitySpecifications
}
\description{
A list containing specifications for the CalculateCarbonIntensity module.
}
\keyword{datasets}
