% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCarbonIntensity.R
\name{DataPresent}
\alias{DataPresent}
\title{Determine whether a list component contains data}
\usage{
DataPresent(List, ComponentName)
}
\arguments{
\item{List}{a list that is to be checked for the presence of the component}

\item{ComponentName}{a string that is the name of the component}
}
\value{
A logical value that is FALSE if the named component is not present
in the list or if the value of the component is NA. Otherwise the value is
TRUE.
}
\description{
\code{DataPresent} determines whether a list component contains data.
}
\details{
Function determines whether a list component contains data if it is not NULL
or NA
}
