% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{Initialize}
\alias{Initialize}
\title{Check and adjust fuel and powertrain proportions inputs.}
\usage{
Initialize(L)
}
\arguments{
\item{L}{A list containing data from preprocessing supplied optional input
files returned by the processModuleInputs function. This list has two
components: Errors and Data.}
}
\value{
A list that is the same as the input list with an additional
Warnings component.
}
\description{
\code{Initialize} checks optional fuel and powertrains proportions datasets
to determine whether they each sum to 1, creates error and warning messages,
and makes adjustments if necessary.
}
\details{
This function processes optional user energy and emissions inputs that have
been preprocessed by the processModuleInputs function. It checks datasets
that specify fuel type or powertrain type proportions to determine whether
they sum to 1. If the sum for a dataset differs from 1 by more than 1%, then
the function returns an error message identifying the problem dataset. If the
sum differs from 1 but the difference is 1% or less it is assumed that the
difference is due to rounding errors and function adjusts the proportions so
that they equal 1. In this case, a warning message is returned as well that
the framework will write to the log.
}
