% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCarbonIntensity.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate a value from a series for a specified year}
\usage{
interpolate(Values_, Years_, Year)
}
\arguments{
\item{Values_}{a vector of values that is numeric or may coerced into
numeric values}

\item{Years_}{a numeric or string vector of years corresponding to the vector
of values}

\item{Year}{a numeric or string value for the year to calculate for.}
}
\value{
A numeric value interpolated for the specified year.
}
\description{
\code{interpolate} finds a value for a specified year from a vector of
values and a corresponding vector of years.
}
\details{
Function determines determines the value for a specified year by
interpolation using a vector of values and a corresponding vector of years.
Smooth splines are used for interpolation.
}
