% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCarbonIntensity.R
\name{interpolateDfVals}
\alias{interpolateDfVals}
\title{Calculate a set of interpolated values for a specified year}
\usage{
interpolateDfVals(Vals_df, Year)
}
\arguments{
\item{Vals_df}{a data frame of values that are numeric or may coerced into
numeric values, and also includes a field named 'Year'.}

\item{Year}{a numeric or string value for the year to calculate for.}
}
\value{
A vector of numeric values interpolated for the specified year.
}
\description{
\code{interpolateDfVals} calculates a set of interpolated values, one for
each field in the data frame.
}
\details{
Function interpolates values for each field in a data frame.
}
