% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignHhVehiclePowertrain.R
\docType{data}
\name{PhevElecProp_ls}
\alias{PhevElecProp_ls}
\title{PHEV electric travel proportion lookup tables and lookup function}
\format{A list of lookup table matrices and a lookup function
\describe{
  \item{Metro_VmRg}{a matrix of electric power proportions by DVMT and battery range for metropolitan areas},
  \item{NonMetro_VmRg}{a matrix of electric power proportions by DVMT and battery range for non-metropolitan areas}
  \item{getElecProp}{a function to get a value from a lookup table}
}}
\source{
CalculateHhEnergyAndEmissions.R script.
}
\usage{
PhevElecProp_ls
}
\description{
Lookup metropolitan and non-metropolitan tables identifying the proportion of
travel by vehicle DVMT and vehicle battery range and a function for
extracting values from the tables.
}
\keyword{datasets}
