% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDefaultValues.R
\docType{data}
\name{PowertrainFuelDefaults_ls}
\alias{PowertrainFuelDefaults_ls}
\title{Default energy and emissions data}
\format{A list containing 17 data frames:
\describe{
  \item{CongestionEfficiency_df}{the relative efficiency of vehicle types in congestion by year}
    \item{LdIce}{congestion efficiency of light-duty internal combustion engine vehicles}
    \item{LdHev}{congestion efficiency of light-duty hybrid-electric engine vehicles}
    \item{LdEv}{congestion efficiency of light-duty battery electric vehicles}
    \item{LdFcv}{congestion efficiency of light-duty fuel cell vehicles}
    \item{HdIce}{congestion efficiency of heavy-duty internal combustion engine vehicles}
  \item{CarbonIntensity_df}{a data frame of the carbon intensities of vehicle energy sources by year}
    \item{Year}{calendar year}
    \item{Gasoline}{carbon intensity of gasoline}
    \item{Diesel}{carbon intensity of diesel}
    \item{Cng}{carbon intensity of compressed natural gas}
    \item{Lng}{carbon intensity of liquified natural gas}
    \item{Ethanol}{carbon intensity of ethanol}
    \item{Biodiesel}{carbon intensity of biodiesel}
    \item{Rng}{carbon intensity of renewable natural gas}
    \item{Electricity}{carbon intensity of electricity}
  \item{CarSvcFuel_df}{a data frame of fuel proportions by fuel type used by car service vehicles by year}
    \item{Year}{calendar year}
    \item{AutoPropGasoline}{gasoline proportion auto fuel}
    \item{AutoPropDiesel}{diesel proportion of auto fuel}
    \item{AutoPropCng}{compressed natural gas proportion of auto fuel}
    \item{LtTrkPropGasoline}{gasoline proportion of light truck fuel}
    \item{LtTrkPropDiesel}{diesel proportion of light truck fuel}
    \item{LtTrkPropCng}{compressed natural gas proportion of light truck fuel}
  \item{ComSvcFuel_df}{a data frame of fuel proportions by fuel type used by commercial service vehicles by year}
    \item{Year}{calendar year}
    \item{AutoPropGasoline}{gasoline proportion auto fuel}
    \item{AutoPropDiesel}{diesel proportion of auto fuel}
    \item{AutoPropCng}{compressed natural gas proportion of auto fuel}
    \item{LtTrkPropGasoline}{gasoline proportion of light truck fuel}
    \item{LtTrkPropDiesel}{diesel proportion of light truck fuel}
    \item{LtTrkPropCng}{compressed natural gas proportion of light truck fuel}
  \item{HhFuel_df}{a data frame of fuel proportions by fuel type used by household vehicles by year}
    \item{Year}{calendar year}
    \item{AutoPropGasoline}{gasoline proportion auto fuel}
    \item{AutoPropDiesel}{diesel proportion of auto fuel}
    \item{AutoPropCng}{compressed natural gas proportion of auto fuel}
    \item{LtTrkPropGasoline}{gasoline proportion of light truck fuel}
    \item{LtTrkPropDiesel}{diesel proportion of light truck fuel}
    \item{LtTrkPropCng}{compressed natural gas proportion of light truck fuel}
  \item{LdvBiofuelMix_df}{a data frame of biofuel proportions for light-duty vehicle fuels by year}
    \item{Year}{calendar year}
    \item{EthanolPropGasoline}{ethanol proportion of gasoline}
    \item{BiodieselPropDiesel}{biodiesel proportion of diesel}
    \item{RngPropCng}{renewable natural gas proportion of compressed natural gas}
  \item{LdvPowertrainCharacteristics_df}{a data frame of light-duty vehicle powertrain characteristics by vehicle model year}
    \item{ModelYear}{Model year}
    \item{AutoIcevMpg}{Average fuel economy of automobiles having internal combustion engines (miles per gallon)}
    \item{LtTrkIcevMpg}{Average fuel economy of light trucks having internal combustion engines (miles per gallon)}
    \item{AutoHevMpg}{Average fuel economy of automobiles having hybrid electric powertrains (miles per gallon)}
    \item{LtTrkHevMpg}{Average fuel economy of light trucks having hybrid electric powertrains (miles per gallon)}
    \item{AutoPhevMpg}{Average fuel economy of automobiles having plug-in hybrid electric powertrains (miles per gallon)}
    \item{AutoPhevMpkwh}{Average power economy of automobiles having plug-in hybrid electric powertrains (miles per kilowatt hour)}
    \item{AutoPhevRange}{Average battery range of automobiles having plug-in hybrid electric powertrains (miles)}
    \item{LtTrkPhevMpg}{Average fuel economy of light trucks having plug-in hybrid electric powertrains (miles per gallon)}
    \item{LtTrkPhevMpkwh}{Average power economy of light trucks having plug-in hybrid electric powertrains (miles per kilowatt hour)}
    \item{LtTrkPhevRange}{Average battery range of light trucks having plug-in hybrid electric powertrains (miles)}
    \item{AutoBevMpkwh}{Average power economy of automobiles having battery electric powertrains (miles per kilowatt hour)}
    \item{AutoBevRange}{Average battery range of automobiles having battery electric powertrains (miles)}
    \item{LtTrkBevMpkwh}{Average power economy of light trucks having battery electric powertrains (miles per kilowatt hour)}
    \item{LtTrkBevRange}{Average battery range of light trucks having battery electric powertrains (miles)}
  \item{HhPowertrain_df}{a data frame of household vehicle powertrain proportions by vehicle model year}
    \item{ModelYear}{Model year}
    \item{AutoPropIcev}{Proportion of automobiles that have internal combustion engine powertrains}
    \item{AutoPropHev}{Proportion of automobiles that have hybrid electric powertrains}
    \item{AutoPropPhev}{Proportion of automobiles that have plug-in hybrid electric powertrains}
    \item{AutoPropBev}{Proportion of automobiles that have battery electric powertrains}
    \item{LtTrkPropIcev}{Proportion of light trucks that have internal combustion engine powertrains}
    \item{LtTrkPropHev}{Proportion of light trucks that have hybrid electric powertrains}
    \item{LtTrkPropPhev}{Proportion of light trucks that have plug-in hybrid electric powertrains}
    \item{LtTrkPropBev}{Proportion of light trucks that have battery electric powertrains}
  \item{CarSvcPowertrain_df}{a data frame of powertrain proportions of car service vehicles by year}
    \item{Year}{calendar year}
    \item{AutoPropIcev}{Proportion of automobiles that have internal combustion engine powertrains}
    \item{AutoPropHev}{Proportion of automobiles that have hybrid electric powertrains}
    \item{AutoPropBev}{Proportion of automobiles that have battery electric powertrains}
    \item{LtTrkPropIcev}{Proportion of light trucks that have internal combustion engine powertrains}
    \item{LtTrkPropHev}{Proportion of light trucks that have hybrid electric powertrains}
    \item{LtTrkPropBev}{Proportion of light trucks that have battery electric powertrains}
  \item{ComSvcPowertrain_df}{a data frame of powertrain proportions of commercial service vehicles by year}
    \item{Year}{calendar year}
    \item{AutoPropIcev}{Proportion of automobiles that have internal combustion engine powertrains}
    \item{AutoPropHev}{Proportion of automobiles that have hybrid electric powertrains}
    \item{AutoPropBev}{Proportion of automobiles that have battery electric powertrains}
    \item{LtTrkPropIcev}{Proportion of light trucks that have internal combustion engine powertrains}
    \item{LtTrkPropHev}{Proportion of light trucks that have hybrid electric powertrains}
    \item{LtTrkPropBev}{Proportion of light trucks that have battery electric powertrains}
  \item{HvyTrkFuel_df}{a data frame of fuel proportions by fuel type used by heavy trucks by year}
    \item{Year}{calendar year}
    \item{PropDiesel}{diesel proportion of heavy truck fuel}
    \item{PropGasoline}{gasoline proportion heavy truck fuel}
    \item{PropLng}{liquified natural gas proportion of heavy truck fuel}
  \item{HvyTrkBiofuelMix_df}{a data frame of biofuel proportions for heavy truck fuels by year}
    \item{Year}{calendar year}
    \item{EthanolPropGasoline}{ethanol proportion of gasoline}
    \item{BiodieselPropDiesel}{biodiesel proportion of diesel}
    \item{RngPropLng}{renewable natural gas proportion of liquified natural gas}
  \item{HvyTrkPowertrainCharacteristics_df}{a data frame of heavy truck powertrain characteristics by year}
    \item{Year}{calendar year}
    \item{HvyTrkIcevMpg}{Average fuel economy of heavy trucks having internal combustion engines (miles per gallon)}
    \item{HvyTrkHevMpg}{Average fuel economy of heavy trucks having hybrid electric powertrains (miles per gallon)}
    \item{HvyTrkBevMpkwh}{Average power economy of heavy trucks having battery electric powertrains (miles per kilowatt hour)}
  \item{HvyTrkPowertrain_df}{a data frame of heavy truck powertrain proportions by year}
    \item{Year}{calendar year}
    \item{HvyTrkPropIcev}{Proportion of heavy trucks that have internal combustion engine powertrains}
    \item{HvyTrkPropBev}{Proportion of heavy trucks that have battery electric powertrains}
  \item{TransitFuel_df}{a data frame of fuel proportions by fuel type for each transit vehicle type by year}
    \item{Year}{calendar year}
    \item{VanPropDiesel}{diesel proportion of transit van fuel}
    \item{VanPropGasoline}{gasoline proportion of transit van fuel}
    \item{VanPropCng}{liquified natural gas proportion of transit van fuel}
    \item{BusPropDiesel}{diesel proportion of transit bus fuel}
    \item{BusPropGasoline}{gasoline proportion of transit bus fuel}
    \item{BusPropCng}{liquified natural gas proportion of transit bus fuel}
    \item{RailPropDiesel}{diesel proportion of rail transit fuel}
    \item{RailPropGasoline}{gasoline proportion of rail transit fuel}
  \item{TransitBiofuelMix_df}{a data frame of biofuel proportions for transit fuels by year}
    \item{Year}{calendar year}
    \item{EthanolPropGasoline}{ethanol proportion of gasoline}
    \item{BiodieselPropDiesel}{biodiesel proportion of diesel}
    \item{RngPropLng}{renewable natural gas proportion of liquified natural gas}
  \item{TransitPowertrainCharacteristics_df}{a data frame of transit vehicle powertrain characteristics by year}
    \item{Year}{Calendar year}
    \item{VanIcevMpg}{Average fuel economy of transit vans having internal combustion engines (miles per gallon)}
    \item{VanHevMpg}{Average fuel economy of transit vans having hybrid electric powertrains (miles per gallon)}
    \item{VanBevMpkwh}{Average power economy of transit vans having battery electric powertrains (miles per kilowatt hour)}
    \item{BusIcevMpg}{Average fuel economy of transit buses having internal combustion engines (miles per gallon)}
    \item{BusHevMpg}{Average fuel economy of transit buses having hybrid electric powertrains (miles per gallon)}
    \item{BusBevMpkwh}{Average power economy of transit buses having battery electric powertrains (miles per kilowatt hour)}
    \item{RailIcevMpg}{Average fuel economy of transit rail vehicles having internal combustion engines (miles per gallon)}
    \item{RailHevMpg}{Average fuel economy of transit rail vehicles having hybrid electric powertrains (miles per gallon)}
    \item{RailEvMpkwh}{Average power economy of transit rail vehicles having electric powertrains (miles per kilowatt hour)}
  \item{TransitPowertrain_df}{a data frame of powertrain proportions of transit vehicles by year}
    \item{Year}{calendar year}
    \item{VanPropIcev}{Proportion of transit vans that have internal combustion engine powertrains}
    \item{VanTrkPropHev}{Proportion of transit vans that have hybrid electric powertrains}
    \item{VanTrkPropBev}{Proportion of transit vans that have battery electric powertrains}
    \item{BusPropIcev}{Proportion of transit buses that have internal combustion engine powertrains}
    \item{BusTrkPropHev}{Proportion of transit buses that have hybrid electric powertrains}
    \item{BusTrkPropBev}{Proportion of transit buses that have battery electric powertrains}
    \item{RailPropIcev}{Proportion of rail transit vehicles that have internal combustion engine powertrains}
    \item{RailTrkPropHev}{Proportion of rail transit vehicles that have hybrid electric powertrains}
    \item{RailTrkPropEv}{Proportion of rail transit vehicles that have electric powertrains}
}}
\source{
LoadDefaultValues.R script.
}
\usage{
PowertrainFuelDefaults_ls
}
\description{
A list of datasets containing default assumptions about fuel and electricity
carbon intensities, vehicle fuel mixess, vehicle powertrain mixes, and
vehicle powertrain mixes by transportation and vehicle type.
}
\keyword{datasets}
